♻️ 스크롤 버튼 배경 추가 및 훅 분리
