+++ frontend/src/pages/HowToPair/HowToPair.styles.ts
@@ -7,7 +7,8 @@ export const Layout = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
-  gap: 20rem;
+  gap: 16rem;
+  overflow-x: hidden;
 
   position: relative;
